#124 Установил jest + supertest. Написал тесты для areasRouter

+++ backend/routers/areas.ts
@@ -19,11 +19,9 @@ areasRouter.get('/', auth, async (req, res, next) => {
 
 areasRouter.post('/', auth, permit('admin'), async (req, res, next) => {
   try {
-    const areaData = await Area.create({
-      name: req.body.name,
-    });
+    const area = await Area.create({ name: req.body.name });
 
-    return res.send(areaData);
+    return res.status(201).send({ message: 'Новая область успешно создана!', area });
   } catch (e) {
     if (e instanceof mongoose.Error.ValidationError) {
       return res.status(400).send(e);
@@ -34,15 +32,19 @@ areasRouter.post('/', auth, permit('admin'), async (req, res, next) => {
 });
 
 areasRouter.delete('/:id', auth, permit('admin'), async (req, res, next) => {
+  const _id = req.params.id as string;
+  if (!mongoose.isValidObjectId(_id)) {
+    return res.status(400).send({ error: 'Некорректный id области' });
+  }
+
   try {
-    const _id = req.params.id as string;
     const area = await Area.findOne({ _id });
     const location = await Location.find({ area: _id });
     const cities = await City.find({ area: _id });
     if (!area) {
       return res.status(404).send({ error: 'Область не существует в базе.' });
     } else if (location.length > 0 || cities.length > 0) {
-      return res.status(404).send({ error: 'Область привязана к другим сущностям ! удаление запрещено' });
+      return res.status(409).send({ error: 'Область привязана к другим сущностям! Удаление запрещено.' });
     }
     const result = await Area.deleteOne({ _id });
     return res.send(result);
